refactor: 폴더 구조 정리, main 변경 사항 반영
