+++ src/app/user/setting/page.tsx
@@ -3,6 +3,7 @@ import Sidebar from "@/common/component/Sidebar";
 import ToastProvider from "@/common/component/Toast";
 import { sidebarWrapper } from "@/styles/shared.css";
 import AccountManagement from "@/view/user/setting/AccountManagement";
+import AlarmSetting from "@/view/user/setting/AlarmSetting";
 import MyProfile from "@/view/user/setting/MyProfile";
 import SettingStep from "@/view/user/setting/SettingStep";
 import StudyList from "@/view/user/setting/StudyList";
@@ -22,7 +23,7 @@ const UserSettingPage = () => {
         .with("my-profile", () => <MyProfile />)
         .with("study-setting", () => <StudyList />)
         .with("account-setting", () => <AccountManagement />)
-        .with("alarm-setting", () => <></>)
+        .with("alarm-setting", () => <AlarmSetting />)
         .exhaustive()}
     </main>
   );
