+++ src/view/user/setting/MyProfile/EditForm/index.tsx
@@ -4,7 +4,7 @@ import { Form, FormController } from "@/shared/component/Form";
 import { handleOnChangeMode } from "@/shared/util/form";
 import Card from "@/view/index/Card";
 import SubmitButton from "@/view/index/SubmitButton";
-import { editCardStyle, footerStyle } from "../MyProfile/index.css";
+import { editCardStyle, footerStyle } from "../index.css";
 import { contentStyle, formStyle, labelStyle } from "./index.css";
 import useEditForm from "./useEditForm";
