+++ src/view/user/setting/MyProfile/index.tsx
@@ -1,5 +1,5 @@
-import EditForm from "../EditForm";
 import { headingStyle } from "../index.css";
+import EditForm from "./EditForm";
 import { containerStyle } from "./index.css";
 
 const MyProfile = () => {
